Lift WordsBlock cursor code to WordsBlockCursor
Lift and generalise WordsBlock's draggable cursor feature into WordsBlockCursor.

Add touch drag and drop functionality.
